requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
langchain>=0.0.350
langgraph>=0.0.10
google-generativeai>=0.3.0
httpx>=0.25.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
typing-extensions>=4.8.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.25.0
pytest-mock>=3.12.0
python-jose>=3.3.0
freezegun>=1.2.2 